move cli to src/rail/cli/rail (#171)
* move cli to src/rail/cli/rail

* fix import in tests

* fix one more import

* fix typo
