+++ src/rail/cli/rail/__main__.py
@@ -1,0 +1,5 @@
+# This file must exist with these contents
+from .commands import cli
+
+if __name__ == "__main__":
+    cli()
