+++ src/rail/cli/rail/commands.py
@@ -4,7 +4,7 @@ import pprint
 import yaml
 
 from rail.core import __version__
-from rail.cli import options, scripts
+from rail.cli.rail import options, scripts
 from rail.interfaces.pz_factory import PZFactory
 from rail.interfaces.tool_factory import ToolFactory
 from rail.utils import catalog_utils
