+++ tests/cli/test_scripts.py
@@ -1,8 +1,8 @@
 import glob
 import os
 import pytest
-from rail.cli import scripts
-from rail.cli.options import args_to_dict, GitMode
+from rail.cli.rail import scripts
+from rail.cli.rail.options import args_to_dict, GitMode
